reorganize, update header areas
