--- frontend/src/components/banner/Slim.tsx
@@ -1,8 +1,0 @@
-const Slim = () => {
-  return (
-    <div className="banner">
-      <div data-cy="banner-title" className="container h-32 md:h-28" />
-    </div>
-  );
-};
-export default Slim;
