+++ frontend/src/components/logo/AlphaLogo.tsx
